chore: fix clippy lints & check fmt
This addresses various warnings from clippy and makes it so that CI and
`just ok` fails if there are formatting issues to help ensure
consistency and clean code.

+++ raylib/src/core/mod.rs
@@ -203,7 +203,7 @@ impl RaylibBuilder {
         }
 
         unsafe {
-            ffi::SetConfigFlags(flags as u32);
+            ffi::SetConfigFlags(flags);
         }
 
         unsafe {
